cracklib: return built-in cracklib dictionary when queried
It's sometimes usuful to know the dictionary used by the cracklib
invocation hidden in libpwquality (for example to gracefully disable
such checks in some environments that lack the dictionary file
automatically). Right now, this is not queriable however, and
libpwquality only allows to query the setting configured in libpwquality
itself. Let's fix that and expose the cracklib built-in path, too, if
nothing was explicitly configured otherwise.